@@ -1,10 +1,12 @@
 buildscript {
+  // ext.kotlin_version = '1.8.22'
+
   repositories {
     mavenCentral()
     google()
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:8.4.0'
+    classpath 'com.android.tools.build:gradle:8.11.0'
   }
 }
 
@@ -41,10 +43,10 @@ keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 android {
   namespace 'sk.breviar.android'
 
-  compileSdkVersion 35
+  compileSdkVersion 36
   
   defaultConfig {
-    targetSdkVersion 35
+    targetSdkVersion 36
       // TODO: api 33 has changes in notifications: POST_NOTIFICATIONS permission
     minSdkVersion 21
     ndk {
@@ -101,9 +103,16 @@ android {
 }
 
 dependencies {
+  /*
   implementation 'com.android.support:support-v4:27.0.0'
   implementation 'com.android.support:support-core-utils:27.0.0'
   implementation 'com.android.support:design:27.0.0'
   implementation 'com.android.support:appcompat-v7:27.0.0'
   implementation 'com.android.support:support-media-compat:27.0.0'
+  */
+  implementation 'androidx.core:core:1.17.0-beta01'
+  implementation 'androidx.appcompat:appcompat:1.7.1'
+  implementation 'androidx.appcompat:appcompat-resources:1.7.1'
+  implementation 'androidx.media:media:1.7.0'
+  implementation 'com.google.android.material:material:1.12.0'
 }
