@@ -2,8 +2,8 @@ package sk.breviar.android;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.widget.CompoundButton;
 
 import sk.breviar.android.BreviarApp;
